Renaming 'constants' to 'docLinks'
