+++ components/raven_tools/sources/keyword-added/keyword-added.ts
@@ -1,6 +1,6 @@
 import { defineSource } from "@pipedream/types";
 import { ConfigurationError } from "@pipedream/platform";
-import { DOCS } from "../../common/constants";
+import { DOCS } from "../../common/docLinks";
 import common from "../common";
 import {
   Keyword,
